+++ ui/app/src/hooks/useInitialize.ts
@@ -12,6 +12,7 @@ import {
 import { getTokenIconUrl } from "@/utils/getTokenIconUrl";
 import { getTokenContract } from "../../../core/src/services/EthbridgeService/tokenContract";
 import { convertImageUrlToDataUrl } from "@/utils/convertImageUrlToDataUrl";
+import { useChains } from "./useChains";
 
 const mirrorToCore = (network: Network) => {
   const data = accountStore.state[network];
@@ -57,8 +58,10 @@ let connectAll = () => {
 export function useInitialize() {
   connectAll();
   connectAll = () => {};
+
   const { usecases, store, services, config } = useCore();
-  // Initialize usecases / watches
+
+  services.ethbridge.addEthereumAddressToPeggyCompatibleCosmosAssets();
 
   services.ethbridge
     .addEthereumAddressToPeggyCompatibleCosmosAssets()
@@ -66,64 +69,57 @@ export function useInitialize() {
       async function generateUniswapWhitelist() {
         const whitelist = {
           name: "Sifchain",
-          logoURI: await convertImageUrlToDataUrl(
-            getTokenIconUrl(Asset("rowan"), window.location.origin)!,
-          ),
+          logoURI: getTokenIconUrl(
+            Asset("rowan"),
+            `https://dex-sifchain-finance.ipns.dweb.link/`,
+          )?.replace("/public/", ""),
           keywords: ["peggy", "pegged assets", "cosmos ecosystem"],
-          tags: {
-            sifchain: {
-              name: "Sifchain",
-              description: "Tokens from the world's first omni-chain DEX",
-            },
-            peggy: {
-              name: "Peggy Bridge",
-              description:
-                "Tokens from Cosmos Ecosystem that have been pegged across Peggy bridge",
-            },
-            cosmos: {
-              name: "Cosmos Ecosystem",
-              description: "Tokens which utilize the Cosmos-SDK",
-            },
-          },
+          tags: {},
           timestamp: new Date().toISOString(),
           tokens: [
-            ...(await Promise.all(
-              [...config.peggyCompatibleCosmosBaseDenoms]
-                .map(async (denom) => {
-                  const web3 = new services.Web3(await getMetamaskProvider());
-                  const asset = config.assets.find(
-                    (a) => a.network === Network.ETHEREUM && a.symbol === denom,
-                  );
-                  if (!asset) return;
-                  const addressOfToken = await services.ethbridge.fetchTokenAddress(
-                    asset,
-                  );
-                  const tokenContract = new web3.eth.Contract(
-                    await fetch(
-                      `https://gist.githubusercontent.com/veox/8800debbf56e24718f9f483e1e40c35c/raw/f853187315486225002ba56e5283c1dba0556e6f/erc20.abi.json`,
-                    ).then((r) => r.json()),
-                    addressOfToken || "",
-                  );
-                  const symbol = await tokenContract.methods.symbol().call();
-                  const decimals = await tokenContract.methods
-                    .decimals()
-                    .call();
-                  const name = await tokenContract.methods.name().call();
-                  const imageUrl = getTokenIconUrl(asset);
-                  if (!imageUrl) return;
-                  const item = {
-                    chainId: 1,
-                    address: addressOfToken,
-                    symbol,
-                    name,
-                    decimals: +decimals,
-                    logoURI: await convertImageUrlToDataUrl(imageUrl),
-                    tags: ["sifchain", "peggy", "cosmos"],
-                  };
-                  return item;
-                })
-                .filter((a) => !!a),
-            )),
+            ...(
+              await Promise.all(
+                [...config.peggyCompatibleCosmosBaseDenoms].map(
+                  async (denom) => {
+                    const web3 = new services.Web3(await getMetamaskProvider());
+                    const asset = config.assets.find(
+                      (a) =>
+                        a.network === Network.ETHEREUM && a.symbol === denom,
+                    );
+                    if (!asset) return;
+                    const addressOfToken = await services.ethbridge.fetchTokenAddress(
+                      asset,
+                    );
+                    const tokenContract = new web3.eth.Contract(
+                      await fetch(
+                        `https://gist.githubusercontent.com/veox/8800debbf56e24718f9f483e1e40c35c/raw/f853187315486225002ba56e5283c1dba0556e6f/erc20.abi.json`,
+                      ).then((r) => r.json()),
+                      addressOfToken || "",
+                    );
+                    const symbol = await tokenContract.methods.symbol().call();
+                    const decimals = await tokenContract.methods
+                      .decimals()
+                      .call();
+                    const name = await tokenContract.methods.name().call();
+                    const imageUrl = getTokenIconUrl(
+                      asset,
+                      "https://dex-sifchain-finance.ipns.dweb.link/",
+                    )?.replace("/public/", "");
+                    if (!imageUrl) return;
+                    const item = {
+                      chainId: 1,
+                      address: addressOfToken,
+                      symbol,
+                      name,
+                      decimals: +decimals,
+                      tags: [],
+                      logoURI: imageUrl,
+                    };
+                    return item;
+                  },
+                ),
+              )
+            ).filter((a) => !!a),
           ],
           version: {
             major: 1,
@@ -146,23 +142,24 @@ export function useInitialize() {
   // Support legacy code that uses sif service getState().
   watch(
     accountStore.state.sifchain,
-    () => {
-      const storeState = accountStore.state.sifchain;
+    (storeState) => {
       const state = services.sif.getState();
+
+      if (storeState.connected && !state.connected) {
+        accountStore.updateBalances(Network.SIFCHAIN);
+        accountStore.pollBalances(Network.SIFCHAIN);
+      }
+
       state.balances = storeState.balances;
       state.address = storeState.address;
       state.connected = storeState.connected;
       state.accounts = [storeState.address];
-
-      if (state.connected) {
-        accountStore.pollBalances(Network.SIFCHAIN);
-      }
     },
     { deep: true },
   );
 
-  // Connect to networks in sequence, starting with Sifchain.
   for (const network of Object.values(Network)) {
+    if (useChains().get(network).chainConfig.hidden) continue;
     accountStore.actions.loadIfConnected(network);
     watch(
       accountStore.refs[network].computed(),